pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-browser-automation"
version = "0.1.0"
description = "AI-powered browser automation tool with natural language control"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "playwright>=1.40",
    "selenium>=4.15",
    "browser-use>=0.1",
    "openai>=1.0",
    "google-generativeai>=0.3",
    "boto3>=1.34",
    "httpx>=0.25",
    "cryptography>=41.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "structlog>=23.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "pytest-xdist>=3.0",
    # Linting & Formatting
    "ruff>=0.4",
    "mypy>=1.8",
    "black>=24.0",
    "isort>=5.13",
    # Security
    "bandit>=1.7",
    "pip-audit>=2.7",
    "detect-secrets>=1.4",
    # Dependency management
    "pip-tools>=7.0",
]

[project.scripts]
ai-browser = "ai_browser_automation.main:main"

# ============================================================
# Ruff — Linting & Style (thay thế flake8, pylint, pycodestyle)
# ============================================================
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["ai_browser_automation", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "DTZ",  # flake8-datetimez
    "T20",  # flake8-print (no print statements)
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate (commented-out code)
    "RUF",  # ruff-specific rules
    "ASYNC",# flake8-async
]
ignore = [
    "S101",  # assert allowed in tests
    "B008",  # function call in default argument (needed for pydantic)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # assert allowed in tests
    "ARG",    # unused arguments OK in fixtures
    "S106",   # hardcoded passwords OK in tests
]

[tool.ruff.lint.isort]
known-first-party = ["ai_browser_automation"]
force-sort-within-sections = true

# ============================================================
# Black — Code Formatting
# ============================================================
[tool.black]
line-length = 100
target-version = ["py310"]

# ============================================================
# isort — Import Sorting
# ============================================================
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["ai_browser_automation"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

# ============================================================
# MyPy — Type Checking
# ============================================================
# ============================================================
# Pytest
# ============================================================
[tool.pytest.ini_options]
asyncio_mode = "auto"

# ============================================================
# MyPy — Type Checking
# ============================================================
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
